+++ src/Zope2/App/startup.py
@@ -20,6 +20,7 @@ from Acquisition import aq_acquire
 from Acquisition import aq_base
 from Acquisition import aq_inner
 from Acquisition import aq_parent
+from Acquisition.interfaces import IAcquirer
 from App.config import getConfiguration
 from time import asctime
 from types import StringType, ListType
@@ -208,6 +209,8 @@ class ZPublisherExceptionHook:
             # Zope 3 uses as well.
             view = queryMultiAdapter((v, REQUEST), name=u'index.html')
             if view is not None:
+                if IAcquirer.providedBy(published):
+                    view = view.__of__(published)
                 v = view()
                 response = REQUEST.RESPONSE
                 response.setStatus(t)